### 2018-10-15 #1024 @whedon
[PRE REVIEW]: Altair: Interactive Statistical Visualizations for Python
**Submitting author:** @domoritz (<a href="http://orcid.org/0000-0002-3110-1053">ROBYN DOMINIK MORITZ</a>)
**Repository:** <a href="https://github.com/altair-viz/altair" target ="_blank">https://github.com/altair-viz/altair</a>
**Version:** v2
