### 2018-10-16 #1024 @domoritz
@arfon Done in https://github.com/altair-viz/altair/pull/1167.

I don't remember who said it but we had an editor volunteer to review the Altair submission. I hope that they see it and speak up. (https://github.com/altair-viz/altair/pull/394#issuecomment-429906564)
