### 2018-10-16 #1024 @arfon
> @arfon Done in altair-viz/altair#1167.

Great. Can we get that merged in?
