### 2018-10-16 #1024 @domoritz
That's a very odd citation style.
