### 2018-10-16 #1024 @ellisonbg
Merged the altair PR.
