### 2018-10-17 #1024 @arfon
> That's a very odd citation style.

I think it's because the `author` string is in the format author AND author and author (e.g. https://github.com/altair-viz/altair/blob/master/paper/paper.bib#L19) rather than Lastname, Firstname and Lastname, Firstname and Lastname, Firstname
