### 2018-10-17 #1024 @domoritz
I fixed the author format in https://github.com/altair-viz/altair/pull/1169
