### 2018-10-18 #1024 @jedbrown
It's somewhat outside my usual wheelhouse, but I have some potential reviewers in mind and would be happy to edit this submission if no other editor emerges to claim it.
