### 2018-10-18 #1024 @arfon
This PR seems to fix the formatting @domoritz : https://github.com/altair-viz/altair/pull/1171

Check out this version I compiled locally based on these changes:
[10.21105.joss.01024.pdf](https://github.com/openjournals/joss-reviews/files/2491965/10.21105.joss.01024.pdf)
