### 2018-10-18 a4faf5a
Paper: clickable link to contributors
