### 2018-10-18 #1024 @jedbrown
Looks good apart from capitalization in the CHI reference.

You can do a basic compile directly using pandoc, but if you want complete decoration you'd install https://github.com/openjournals/whedon/ (which has a lot of Ruby dependencies and ultimately calls pandoc with many arguments).
`pandoc -o paper.pdf --filter pandoc-citeproc --csl /path/to/apa.csl paper.md`
