### 2018-10-18 16ee16a
Paper: Capitalization in references
