+++ paper/paper.bib
@@ -1,7 +1,7 @@
 @article{2011-d3,
  title = {{D3}: Data-Driven Documents},
  author = {Bostock, Michael and Ogievetsky, Vadim and Heer, Jeffrey},
- journal = {{IEEE Trans. Visualization \& Computer Graphics (Proc. InfoVis)}},
+ journal = {{IEEE Trans. Visualization \& Comp. Graphics (Proc. InfoVis)}},
  year = {2011},
  url = {http://idl.cs.washington.edu/papers/d3},
 }
@@ -9,7 +9,7 @@
 @article{2017-vega-lite,
  title = {Vega-Lite: A Grammar of Interactive Graphics},
  author = {Satyanarayan, Arvind and Moritz, Dominik and Wongsuphasawat, Kanit and Heer, Jeffrey},
- journal = {{IEEE Trans. Visualization \& Computer Graphics (Proc. InfoVis)}},
+ journal = {{IEEE Trans. Visualization \& Comp. Graphics (Proc. InfoVis)}},
  year = {2017},
  url = {http://idl.cs.washington.edu/papers/vega-lite},
 }
@@ -25,7 +25,7 @@
 @article{2016-voyager,
  title = {Voyager: Exploratory Analysis via Faceted Browsing of Visualization Recommendations},
  author = {Wongsuphasawat, Kanit and Moritz, Dominik and Anand, Anushka and Mackinlay, Jock and Howe, Bill and Heer, Jeffrey},
- journal = {{IEEE Trans. Visualization \& Computer Graphics (Proc. InfoVis)}},
+ journal = {{IEEE Trans. Visualization \& Comp. Graphics (Proc. InfoVis)}},
  year = {2016},
  url = {http://idl.cs.washington.edu/papers/voyager},
 }
@@ -33,7 +33,7 @@
 @article{2016-reactive-vega-architecture,
  title = {Reactive Vega: A Streaming Dataflow Architecture for Declarative Interactive Visualization},
  author = {Satyanarayan, Arvind and Russell, Ryan and Hoffswell, Jane and Heer, Jeffrey},
- journal = {{IEEE Trans. Visualization \& Computer Graphics (Proc. InfoVis)}},
+ journal = {{IEEE Trans. Visualization \& Comp. Graphics (Proc. InfoVis)}},
  year = {2016},
  url = {http://idl.cs.washington.edu/papers/reactive-vega-architecture},
 }
