### 2018-10-18 835c768
Paper: Capitalize CHI
