+++ paper/paper.bib
@@ -17,7 +17,7 @@
 @inproceedings{2017-voyager2,
  title = {Voyager 2: Augmenting Visual Analysis with Partial View Specifications},
  author = {Wongsuphasawat, Kanit and Qu, Zening and Moritz, Dominik and Chang, Riley and Ouk, Felix and Anand, Anushka and Mackinlay, Jock and Howe, Bill and Heer, Jeffrey},
- booktitle = {ACM Human Factors in Computing Systems (CHI)},
+ booktitle = {ACM Human Factors in Computing Systems ({CHI})},
  year = {2017},
  url = {http://idl.cs.washington.edu/papers/voyager2},
 }
