### 2018-10-18 #1024 @domoritz
I fixed the capitalization of CHI in https://github.com/altair-viz/altair/pull/1174.
