### 2018-10-19 87743f9
Paper: Remove word
