+++ paper/paper.md
@@ -97,6 +97,6 @@ on the bar chart. The selections drive filters in the other plot. The code for t
 # Acknowledgements
 
 We thank the many contributors that created examples, wrote documentation, and reported bugs. You can find [an up-to-date
-list of contributors is on GitHub](https://github.com/altair-viz/altair/graphs/contributors).
+list of contributors on GitHub](https://github.com/altair-viz/altair/graphs/contributors).
 
 # References
