### 2018-10-27 a4010bf
update troubleshooting guide for empty charts caused by adblocker
