### 2018-11-01 #1057 @whedon
[REVIEW]: Altair: Interactive Statistical Visualizations for Python
**Submitting author:** @domoritz (<a href="http://orcid.org/0000-0002-3110-1053">ROBYN DOMINIK MORITZ</a>)
**Repository:** <a href="https://github.com/altair-viz/altair" target ="_blank">https://github.com/altair-viz/altair</a>
**Version:** v2.3.0

- [x] **Version:** Does the release version given match the GitHub release (v2)?

- [x] **Version:** Does the release version given match the GitHub release (v2)?

- [x] **Version:** Does the release version given match the GitHub release (v2)?
