### 2018-11-01 #1057 @jedbrown
@dnszafir @terrytangyuan @kellieotto :wave: Welcome! The comments from whedon above outline the review process. I'll be watching this thread if you have any questions. Thanks!
