### 2018-11-01 71c2bda
"typing" package apparently creates issues if installed on Python3.6+
see  https://ci.appveyor.com/project/ericpre/hyperspy-bundle/builds/19660402/job/gs9l31okjhplckur
