+++ requirements.txt
@@ -5,4 +5,4 @@ numpy
 pandas
 six
 toolz
-typing
+typing;python_version<"3.5"
