### 2018-11-01 #1024 @jedbrown
@adavidzh If you put it that way, I'm happy to include you. I figured we had sufficient diversity of reviewers and this way we wouldn't feel bad about pinging you in the event of a subsequent submission relevant to your interests in the next coming months.
