### 2018-11-01 #1024 @adavidzh
@jedbrown I'm happy to stay fresh for something else. It's just that the follow-up was not clear to me and I was perhaps over-eager for my first time. Thanks!
