### 2018-11-11 6b47b92
Used url data import for vega_datasets that are not packaged with the library
