+++ altair/vegalite/v2/examples/simple_histogram.py
@@ -8,7 +8,7 @@ https://vega.github.io/vega-lite/examples/histogram.html
 import altair as alt
 from vega_datasets import data
 
-source = data.movies()
+source = data.movies.url
 
 alt.Chart(source).mark_bar().encode(
     alt.X("IMDB_Rating:Q", bin=True),
