+++ altair/vegalite/v2/examples/us_state_capitals.py
@@ -9,7 +9,7 @@ import altair as alt
 from vega_datasets import data
 
 states = alt.topo_feature(data.us_10m.url, 'states')
-capitals = data.us_state_capitals()
+capitals = data.us_state_capitals.url
 
 # US states background
 background = alt.Chart(states).mark_geoshape(
