+++ altair/vegalite/v2/examples/choropleth.py
@@ -8,13 +8,13 @@ import altair as alt
 from vega_datasets import data
 
 counties = alt.topo_feature(data.us_10m.url, 'counties')
-unemp_data = data.unemployment()
+source = data.unemployment.url
 
 alt.Chart(counties).mark_geoshape().encode(
     color='rate:Q'
 ).transform_lookup(
     lookup='id',
-    from_=alt.LookupData(unemp_data, 'id', ['rate'])
+    from_=alt.LookupData(source, 'id', ['rate'])
 ).project(
     type='albersUsa'
 ).properties(
