+++ altair/vegalite/v2/examples/histogram_with_a_global_mean_overlay.py
@@ -7,17 +7,17 @@ This example shows a histogram with a global mean overlay.
 import altair as alt
 from vega_datasets import data
 
-source = data.movies()
+source = data.movies.url
 
 base = alt.Chart(source)
 
 bar = base.mark_bar().encode(
-    x=alt.X('IMDB_Rating', bin=True, axis=None),
+    x=alt.X('IMDB_Rating:Q', bin=True, axis=None),
     y='count()'
 )
 
 rule = base.mark_rule(color='red').encode(
-    x='mean(IMDB_Rating)',
+    x='mean(IMDB_Rating):Q',
     size=alt.value(5)
 )
 
