### 2018-11-12 #1057 @terrytangyuan
Great software and nice summarization of the background in the paper. A couple of questions and suggestions:

* You provided a section [here](https://github.com/altair-viz/altair#feedback-and-contribution) in `README.md` that mentions how to give feedback and contribute. Could you put it in a `CONTRIBUTING.md` file?
* Just want to double check with the other authors: has @domoritz (the submitting author) made significant contributions to this software? It doesn't look like so from the commit history [here](https://github.com/altair-viz/altair/commits?author=domoritz) since most of the commits are only related to this particular paper. However, @domoritz might have contributed significantly offline during the collaboration with the UW Interactive Data Lab (I noticed that he's part of the lab). Maybe @jakevdp and @ellisonbg can chime in and confirm?
* Does any of the references have DOI? If so, would you please add them?
