### 2018-11-12 #1057 @domoritz
Thank you for the review @terrytangyuan!

I'm the co-author of Vega-Lite, created the schema that Altair is generated from, and maintain the Vega-Lite integration in Jupyter Notebooks and Jupyterlab. I created the submission because I did the last push of the paper in https://github.com/altair-viz/altair/pull/394.

I'll send PRs for the other two improvements.
