### 2018-11-12 3cbea70
[Paper] Add DOIs
