+++ paper/paper.bib
@@ -4,6 +4,7 @@
  journal = {IEEE Trans. Visualization \& Comp. Graphics (Proc. InfoVis)},
  year = {2011},
  url = {http://idl.cs.washington.edu/papers/d3},
+ doi = {10.1109/TVCG.2011.185},
 }
 
 @article{2017-vega-lite,
@@ -12,6 +13,7 @@
  journal = {IEEE Trans. Visualization \& Comp. Graphics (Proc. InfoVis)},
  year = {2017},
  url = {http://idl.cs.washington.edu/papers/vega-lite},
+ doi = {10.1109/TVCG.2016.2599030},
 }
 
 @inproceedings{2017-voyager2,
@@ -20,6 +22,7 @@
  booktitle = {ACM Human Factors in Computing Systems ({CHI})},
  year = {2017},
  url = {http://idl.cs.washington.edu/papers/voyager2},
+ doi = {10.1145/3025453.3025768},
 }
 
 @article{2016-voyager,
@@ -28,6 +31,7 @@
  journal = {IEEE Trans. Visualization \& Comp. Graphics (Proc. InfoVis)},
  year = {2016},
  url = {http://idl.cs.washington.edu/papers/voyager},
+ doi = {10.1109/TVCG.2015.2467191},
 }
 
 @article{2016-reactive-vega-architecture,
@@ -36,6 +40,7 @@
  journal = {IEEE Trans. Visualization \& Comp. Graphics (Proc. InfoVis)},
  year = {2016},
  url = {http://idl.cs.washington.edu/papers/reactive-vega-architecture},
+ doi = {10.1109/TVCG.2015.2467091},
 }
 
 @book{1977-exploratory,
@@ -43,7 +48,8 @@
   author={Tukey, John W},
   volume={2},
   year={1977},
-  publisher={Reading, Mass.}
+  publisher={Reading, Mass.},
+  doi = {10.1002/bimj.4710230408 },
 }
 
 @book{2005-grammar,
@@ -53,4 +59,5 @@
  isbn = {0387245448},
  publisher = {Springer-Verlag New York, Inc.},
  address = {Secaucus, NJ, USA},
+ doi = {10.1007/0-387-28695-0},
 }
