### 2018-11-14 #1057 @dnszafir
Agreed on acceptance, @jedbrown. Overall, Altair looks like a fantastic and well-documented resource. I ran into a minor challenge with the installation that may be worth addressing given the broad target audience.

I followed the installation instructions as detailed in the repo (which were clear, simple, and well-articulated). Once installed, I ran into a few issues that turned out to be a stale version of Anaconda (specifically: https://github.com/conda-forge/pandas-feedstock/issues/51). It may be worth noting on the Troubleshooting guide to check for a sufficiently up-to-date version of Anaconda.

After resolving this issue, the plot did not render, but the Troubleshooting guide was partially helpful in resolving the issue. Given the broad target audience, it is worth mentioning that iPython may need to be explicitly updated under "JupyterLab: Textual Chart Representation" as the documented 'Change Kernel' solution did not work in this case. Keep in mind my configuration was probably more out-of-date than the target user as I am not a regular notebook user, but a brief clarification may help others encountering these issues.

As a future thought given the target audience: a growing number of people (both formal students and self-taught) are learning programming and data science through Python. Given the target goals of the package, it might be useful to call out examples that are especially helpful for learning and/or illustrating foundational data science concepts.
