### 2018-11-14 #1057 @domoritz
Thank you for the feedback @dnszafir!

@jakevdp will appreciate the feedback on the getting started guide. The troubleshooting guide could suggest updating the Python installation as a possible step to resolve issues.

Altair is definitely designed with (novice) data scientists in mind. @jakevdp gave a tutorial at PyCon and all the materials are at https://github.com/altair-viz/altair-tutorial. There is also a case study that mirrors the Vega-Lite tutorial on exploring weather data at https://altair-viz.github.io/case_studies/exploring-weather.html. Is that something you had in mind?
