### 2018-11-14 1200ebd
paper.md: small grammar fix
