### 2018-11-14 e077a8a
paper.bib: fix capitalization and book/inproceedings details
