### 2018-11-14 #1057 @jedbrown
Thanks, @kellieotto and @dnszafir!

@domoritz I have a couple minor (mostly bib) edits here: https://github.com/altair-viz/altair/pull/1195
Also, is it intended that the `notebooks/*.ipynb` work? They look unclean (was last executed out of order) and I see some errors. For example,
```python
import altair.vegalite.v2 as vl
vl.vegalite(...)
```
fails because `vl.vegalite` does not exist. Can you look at these?
