### 2018-11-14 #1057 @domoritz
Good catch. I'm fixing the notebooks in https://github.com/altair-viz/altair/pull/1196. The property is called `VegaLite`.
