+++ doc/user_guide/interactions.rst
@@ -35,7 +35,7 @@ to it. Here is a simple scatter-plot created from the ``cars`` dataset:
     import altair as alt
     from vega_datasets import data
 
-    cars = data.cars.url
+    cars = data.cars()
 
     alt.Chart(cars).mark_point().encode(
         x='Miles_per_Gallon:Q',
@@ -152,7 +152,7 @@ selection:
     :output: none
 
     def make_example(selector):
-        cars = data.cars.url
+        cars = data.cars()
 
         return alt.Chart(cars).mark_rect().encode(
             x="Cylinders:O",
