### 2018-11-15 c6749d7
Docs:UserGuide: fix import from vega_datasets
