+++ doc/user_guide/internals.rst
@@ -32,7 +32,7 @@ from which we can output the JSON representation:
     import altair as alt
     from vega_datasets import data
 
-    chart = alt.Chart(data.cars.url).mark_point().encode(
+    chart = alt.Chart(data.cars()).mark_point().encode(
         x='Horsepower:Q',
         y='Miles_per_Gallon:Q',
         color='Origin:N',
