### 2018-11-15 #1057 @jedbrown
Looks good to me now. @domoritz Go ahead and create an archive on Zenodo or similar.
