### 2018-11-16 #1057 @domoritz
I uploaded the current version to Zenodo at https://zenodo.org/record/1489327
