### 2018-11-16 #1057 @jedbrown
Sorry to not be more explicit, @domoritz. We normally ask for the Zenodo archive to be a tagged version of the software (the interface encourages that IIRC) and the author list of the archive should match the author list of the paper. Can you do that?
