### 2018-11-16 #1057 @domoritz
Here is the new version: https://zenodo.org/record/1489372
