### 2018-11-16 #1057 @jedbrown
@arfon This is basically ready to accept, but the archived version (v2.2.2) does not include the paper (it's the most recent release and there are no important changes since). My understanding is that we want a new version (v2.2.3 for example) to be tagged and archived. Can you confirm?
