### 2018-11-16 #1057 @domoritz
@jakevdp Could you tag a new version?
