### 2018-11-17 cf62844
DOC: fix typo in user guide
