### 2018-11-20 9df132a
MAINT: move all examples to top level
