+++ altair/vegalite/v2/__init__.py
@@ -2,8 +2,6 @@
 from .schema import *
 from .api import *
 
-from . import examples
-
 from ...datasets import (
     list_datasets,
     load_dataset
