### 2018-11-20 200fd26
ENH: add renderer set_embed_options() method
