### 2018-11-21 da309c9
Added example of bar chart with negative values
