+++ altair/examples/negative_bar_chart.py
@@ -15,5 +15,5 @@ alt.Chart(source).mark_bar().encode(
         alt.datum.nonfarm_change > 0,
         alt.value("steelblue"),  # The positive color
         alt.value("orange")  # The negative color
-    ),
+    )
 ).properties(width=600)
