### 2018-11-21 c99b560
DOC: Cleaned up simple strip plot example's docs
