### 2018-11-21 ae523b6
DOC: Tidy up line chart with points example, move it down the gallery page
