### 2018-11-21 af3ab40
DOC: Standardized highlighted bar chart examples
