+++ altair/examples/bar_chart_with_highlighted_bar.py
@@ -1,21 +1,21 @@
 """
-Simple Highlighted Bar Chart
-============================
-This example shows a basic bar chart with a single bar singled out for a highlight.
+Bar Chart with Highlighted Bar
+------------------------------
+This example shows a basic bar chart with a single bar highlighted.
 """
 # category: bar charts
 import altair as alt
 from vega_datasets import data
 
-source = data.population.url
+source = data.wheat()
 
 alt.Chart(source).mark_bar().encode(
-    x="year:O",
-    y="sum(people):Q",
+    x='year:O',
+    y="wheat:Q",
     # The highlight will be set on the result of a conditional statement
     color=alt.condition(
-        alt.datum.year == 1970,  # If the year is 1970 this test returns True,
+        alt.datum.year == 1810,  # If the year is 1810 this test returns True,
         alt.value('orange'),     # which sets the bar orange.
         alt.value('steelblue')   # And if it's not true it sets the bar steelblue.
     )
-)
+).properties(width=600)
