### 2018-11-21 63a31e6
Have bar chart use the new vega_datasets source
