+++ altair/examples/bar_chart_with_negatives.py
@@ -5,8 +5,9 @@ This example shows a bar chart with both positive and negative values.
 """
 # category: bar charts
 import altair as alt
+from vega_datasets import data
 
-source = "https://raw.githubusercontent.com/datadesk/vega-datasets/gh-pages/data/us-employment.csv"
+source = data.us_employment()
 
 alt.Chart(source).mark_bar().encode(
     x="month:T",
