### 2018-11-21 7a8f5db
Regroup error bars chart in the example gallery
