### 2018-11-21 62179e1
DOC: Standardize highlighted bar chart examples (#1213)
