### 2018-11-22 15d794a
DOC: Standardized layered bar chart with dual axis
