### 2018-11-22 83425e9
DOC: Standardized bar chart with mean overlay example
