### 2018-11-23 #1057 @jedbrown
@jakevdp Can you comment on this? Other papers I have edited archive a tagged version that contains the post-review repo/paper.
