### 2018-11-24 689d041
DOC: Standardize the horizontal bar chart example
