### 2018-11-24 829161c
DOC: Added case study with Iowa electricity data
