### 2018-11-24 1cd9cdb
Update iowa_electricity.py
