### 2018-11-24 d030544
DOC: Improved barley case study
