### 2018-11-24 4bbdc66
Update beckers_barley_trellis_plot.py
