### 2018-11-24 4d7f58f
Update beckers_barley_trellis_plot.py
