### 2018-11-24 75fbce2
Update beckers_barley_trellis_plot.py
