+++ altair/examples/beckers_barley_trellis_plot.py
@@ -25,6 +25,6 @@ alt.Chart(source, title="The Morris Mistake").mark_point().encode(
     row=alt.Row(
         'site',
         title="",
-        sort=alt.EncodingSortField(field='yield', op='sum', order='descending'),
+        sort=alt.EncodingSortField(field='yield', op='sum', order='descending')
     )
 ).configure_view(stroke="transparent")
