### 2018-11-24 5dae3e9
Update beckers_barley_trellis_plot.py
