### 2018-11-25 4adcb62
DOC: Change the default category to Other Charts
