### 2018-11-26 7141263
DOC: Added more fully developed us_employment case study
