### 2018-11-26 cc10e3d
Update us_employment.py
