+++ altair/examples/us_employment.py
@@ -22,7 +22,6 @@ presidents = alt.pd.DataFrame([
     }
 ])
 
-
 bars = alt.Chart(source, title="The U.S. employment crash during the Great Recession").mark_bar().encode(
     x=alt.X("month:T", axis=alt.Axis(title="")),
     y=alt.Y("nonfarm_change:Q", axis=alt.Axis(title="Change in non-farm employment (in thousands)")),
