### 2018-11-27 3c0b208
DOC: Moved Gantt chart into the other category
