### 2018-11-28 b967533
Add example of custom data transformer to docs
The example shows how the json data transformer can be used to store the
data files in a separate directory.
