### 2018-11-29 16f72c0
DOC: Change grouped bar chart example
