### 2018-11-29 821f3ca
DOC: Remove duplicate us population examples
