### 2018-11-29 5919401
Replace link of pdvega repo
pdvega is no under altair-viz. The old link was not working anymore.
