### 2018-11-29 059bd04
DOC: Simplified version of a grouped bar chart
