### 2018-11-29 e99dbb7
A little more docs
