### 2018-11-29 0a5e754
Both a vertical and horizontal example
