+++ altair/examples/grouped_bar_chart_horizontal.py
@@ -24,7 +24,7 @@ alt.Chart(source).mark_bar().encode(
     # The field used to group the subelements
     row='variety'
 ).configure_view(
-    stroke='transparent'
+    stroke='transparent'  # Remove the trellis frames so multiple charts appear as one.
 ).transform_filter(
     alt.datum.site == 'Morris'
 )
