### 2018-11-30 183e7b1
DOC: Relabel line chart variable
