+++ altair/examples/line_chart_with_points.py
@@ -11,10 +11,10 @@ import pandas as pd
 x = np.arange(100)
 source = pd.DataFrame({
   'x': x,
-  'sin(x)': np.sin(x / 5)}
+  'f(x)': np.sin(x / 5)}
 )
 
 alt.Chart(source).mark_line(point=True).encode(
     x='x',
-    y='sin(x)'
+    y='f(x)'
 )
