### 2018-12-01 eb7f0fe
Update grouped_bar_chart_horizontal.py
