### 2018-12-02 94f5562
Update grouped_bar_chart.py
