+++ altair/examples/grouped_bar_chart.py
@@ -1,7 +1,7 @@
 """
 Grouped Bar Chart
 -----------------
-This example shows a grouped bar chart achieved by making cosmetic changes to a trellis plot.
+This example shows a grouped bar chart.
 """
 # category: bar charts
 import altair as alt
@@ -10,21 +10,8 @@ from vega_datasets import data
 source = data.barley()
 
 alt.Chart(source).mark_bar().encode(
-    # The field used to define the subelements of each group
-    x=alt.X(
-        'year:N',
-        axis=alt.Axis(title="")
-    ),
-    # The length of the bars
-    y=alt.Y(
-        'yield:Q',
-        axis=alt.Axis(grid=False)
-    ),
+    x='year:O',
+    y='sum(yield):Q',
     color='year:N',
-    # The field used to group the subelements
-    column='variety'
-).configure_view(
-    stroke='transparent'  # Remove the trellis frames so multiple charts appear as one.
-).transform_filter(
-    alt.datum.site == 'Morris'
+    column='site:N'
 )
