### 2018-12-04 5437053
MAINT: streamline code generation (no change to user)
