### 2018-12-04 7719713
Bump vega-lite version to 3.0.0-rc8
