+++ altair/vegalite/v3/api.py
@@ -1,0 +1,1669 @@
+# -*- coding: utf-8 -*-
+
+import warnings
+
+import hashlib
+import json
+import jsonschema
+import six
+import pandas as pd
+
+from .schema import core, channels, mixins, Undefined, SCHEMA_URL
+
+from .data import data_transformers, pipe
+from ... import utils, expr
+from .display import renderers, VEGALITE_VERSION, VEGAEMBED_VERSION, VEGA_VERSION
+from .theme import themes
+
+# ------------------------------------------------------------------------
+# Data Utilities
+def _dataset_name(values):
+    """Generate a unique hash of the data
+
+    Parameters
+    ----------
+    values : list or dict
+        A list/dict representation of data values.
+
+    Returns
+    -------
+    name : string
+        A unique name generated from the hash of the values.
+    """
+    if isinstance(values, core.InlineDataset):
+        values = values.to_dict()
+    values_json = json.dumps(values, sort_keys=True)
+    hsh = hashlib.md5(values_json.encode()).hexdigest()
+    return 'data-' + hsh
+
+
+def _consolidate_data(data, context):
+    """If data is specified inline, then move it to context['datasets']
+
+    This function will modify context in-place, and return a new version of data
+    """
+    values = Undefined
+    kwds = {}
+
+    if isinstance(data, core.InlineData):
+        if data.name is Undefined and data.values is not Undefined:
+            values = data.values
+            kwds = {'format': data.format}
+
+    elif isinstance(data, dict):
+        if 'name' not in data and 'values' in data:
+            values = data['values']
+            kwds = {k:v for k,v in data.items() if k != 'values'}
+
+    if values is not Undefined:
+        name = _dataset_name(values)
+        data = core.NamedData(name=name, **kwds)
+        context.setdefault('datasets', {})[name] = values
+
+    return data
+
+
+def _prepare_data(data, context):
+    """Convert input data to data for use within schema
+
+    Parameters
+    ----------
+    data :
+        The input dataset in the form of a DataFrame, dictionary, altair data
+        object, or other type that is recognized by the data transformers.
+    context : dict
+        The to_dict context in which the data is being prepared. This is used
+        to keep track of information that needs to be passed up and down the
+        recursive serialization routine, such as global named datasets.
+    """
+    if data is Undefined:
+        return data
+
+    # convert dataframes to dict
+    if isinstance(data, pd.DataFrame):
+        data = pipe(data, data_transformers.get())
+
+    # convert string input to a URLData
+    if isinstance(data, six.string_types):
+        data = core.UrlData(data)
+
+    # consolidate inline data to top-level datasets
+    if data_transformers.consolidate_datasets:
+        data = _consolidate_data(data, context)
+
+    # if data is still not a recognized type, then return
+    if not isinstance(data, (dict, core.Data, core.UrlData,
+                             core.InlineData, core.NamedData)):
+        warnings.warn("data of type {} not recognized".format(type(data)))
+
+    return data
+
+
+# ------------------------------------------------------------------------
+# Aliases & specializations
+Bin = core.BinParams
+
+@utils.use_signature(core.LookupData)
+class LookupData(core.LookupData):
+    def to_dict(self, *args, **kwargs):
+        """Convert the chart to a dictionary suitable for JSON export"""
+        copy = self.copy(ignore=['data'])
+        context = kwargs.get('context', {})
+        copy.data = _prepare_data(copy.data, context)
+        return super(LookupData, copy).to_dict(*args, **kwargs)
+
+
+@utils.use_signature(core.FacetMapping)
+class FacetMapping(core.FacetMapping):
+    _class_is_valid_at_instantiation = False
+
+    def to_dict(self, *args, **kwargs):
+        copy = self.copy()
[…]
