+++ tools/generate_schema_wrapper.py
@@ -27,6 +27,7 @@ SCHEMA_VERSION = {
     'vega-lite': {
         'v1': 'v1.3.1',
         'v2': 'v2.6.0',
+        'v3': 'v3.0.0-rc8',
     }
 }
 
