+++ altair/vegalite/v3/__init__.py
@@ -1,0 +1,21 @@
+# flake8: noqa
+from .schema import *
+from .api import *
+
+from ...datasets import (
+    list_datasets,
+    load_dataset
+)
+
+from ... import expr
+from ...expr import datum
+
+from .display import VegaLite, renderers
+
+from .data import (
+    MaxRowsError,
+    pipe, curry, limit_rows,
+    sample, to_json, to_csv, to_values,
+    default_data_transformer,
+    data_transformers
+)

[…]
