### 2018-12-04 7e170ec
ENH: set defaultStyle to True by default in HTML output
Note that this also affects the Colab and Kaggle renderers
