### 2018-12-04 d797bc3
TST: update tests for new embed options
