### 2018-12-04 fef3357
MAINT: simplify embed_options defaults
