### 2018-12-04 59a9c2e
MAINT: improve embedOpt testing
