### 2018-12-07 b5a65e3
DOC: use local wheat dataset in examples (requires vega_datasets v0.8
