+++ altair/examples/bar_chart_with_mean_line.py
@@ -7,7 +7,7 @@ This example shows the mean value overlayed on a bar chart.
 import altair as alt
 from vega_datasets import data
 
-source = data.wheat.url
+source = data.wheat()
 
 bar = alt.Chart(source).mark_bar().encode(
     x='year:O',
