### 2018-12-07 117c689
DOC: add class template
