### 2018-12-07 b1c9ecb
MAINT: update Changelog for release 2.3
