### 2018-12-07 193473d
MAINT: bump version to 2.3.0
