+++ doc/conf.py
@@ -74,9 +74,9 @@ author = 'Brian Granger and Jake VanderPlas'
 # built documents.
 #
 # The short X.Y version.
-version = '2.3.0dev0'
+version = '2.3.0'
 # The full version, including alpha/beta/rc tags.
-release = '2.3.0dev0'
+release = '2.3.0'
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
