### 2018-12-07 4ff8aa2
MAINT: bump version to 2.4.0dev0
