+++ doc/conf.py
@@ -74,9 +74,9 @@ author = 'Brian Granger and Jake VanderPlas'
 # built documents.
 #
 # The short X.Y version.
-version = '2.3.0'
+version = '2.4.0dev0'
 # The full version, including alpha/beta/rc tags.
-release = '2.3.0'
+release = '2.4.0dev0'
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
