### 2018-12-07 f7cf187
DOC: Example syntax cleanup
