### 2018-12-07 #1057 @jedbrown
Thanks! In double-checking, I noticed one more messed up citation (comma should be a semicolon). @arfon Is it possible to possible to include this trivial patch in the paper of record without requiring a new tag/archive?

```
diff --git i/paper/paper.md w/paper/paper.md
index 96dc1b4..bc99c65 100644
--- i/paper/paper.md
+++ w/paper/paper.md
@@ -70,7 +70,7 @@ JSON data, which is then rendered in a user-interface such as the Jupyter Notebo
 specification [@2016-reactive-vega-architecture], which is then parsed and executed using a reactive runtime that
 internally makes use of D3.js [@2011-d3].

-The declarative nature of the Vega-Lite visualization grammar [@2005-grammar, @2017-vega-lite], and its encoding in a
+The declarative nature of the Vega-Lite visualization grammar [@2005-grammar; @2017-vega-lite], and its encoding in a
 formal JSON schema, provide Altair with a number of benefits. First, much of the Altair Python code and tests are
 generated from the Vega-Lite JSON schema, ensuring strict conformance with the Vega-Lite specification. Second, the JSON
 data produced by Altair and consumed by Vega-Lite provides a natural serialization and file format for statistical
```
