### 2018-12-07 #1057 @arfon
> Thanks! In double-checking, I noticed one more messed up citation (comma should be a semicolon). @arfon Is it possible to possible to include this trivial patch in the paper of record without requiring a new tag/archive?

Yes, feel free to just update `master` with the fix.
