### 2018-12-07 #1057 @jedbrown
Thanks. @domoritz Can you apply that trivial fix in the repository?
