### 2018-12-07 #1057 @domoritz
@jakevdp can you apply the patch?
