### 2018-12-07 c3c2569
[Paper] Fix reference format
Applying https://github.com/openjournals/joss-reviews/issues/1057#issuecomment-445330891
