### 2018-12-07 #1057 @domoritz
@jakevdp Can you merge https://github.com/altair-viz/altair/pull/1258?
