### 2018-12-07 #1057 @domoritz
Merged.
