### 2018-12-08 f087788
DOC: change axis=Axis(title=title) to title=title
