+++ altair/examples/beckers_barley_trellis_plot.py
@@ -12,14 +12,16 @@ source = data.barley()
 alt.Chart(source, title="The Morris Mistake").mark_point().encode(
     alt.X(
         'yield:Q',
+        title="Barley Yield (bushels/acre)",
         scale=alt.Scale(zero=False),
-        axis=alt.Axis(grid=False, title="Barley Yield (bushels/acre)")
+        axis=alt.Axis(grid=False)
     ),
     alt.Y(
         'variety:N',
+        title="",
         sort=alt.EncodingSortField(field='yield', op='sum', order='descending'),
         scale=alt.Scale(rangeStep=20),
-        axis=alt.Axis(title="", grid=True)
+        axis=alt.Axis(grid=True)
     ),
     color=alt.Color('year:N', legend=alt.Legend(title="Year")),
     row=alt.Row(
