### 2018-12-08 #1057 @arfon
@dnszafir, @terrytangyuan, @kellieotto - many thanks for your reviews here and to @jedbrown for editing this submission ✨

@domoritz @jakevdp - your paper is now accepted into JOSS :zap::rocket::boom:
